% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sd_robust}
\alias{sd_robust}
\title{Robust Standard Deviation}
\usage{
sd_robust(x, p = c(0.1, 0.9))
}
\arguments{
\item{x}{vector of values}

\item{p}{quantile probabilities.}
}
\value{
Standard deviation within selected quantiles = Robust SD
}
\description{
Robust Standard Deviation
}
