% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{staining_index}
\alias{staining_index}
\title{Calculate Staining Index}
\usage{
staining_index(positive, negative, SD_FUN = sd_robust)
}
\arguments{
\item{positive}{Vector of values for positive cells}

\item{negative}{Vector of values for negative cells}

\item{SD_FUN}{Function for calculating Standard Deviation}
}
\description{
((MEDpos - MEDneg) / 2*SD(neg))
}
